Remove all 'withPassword' stuff and make password users default
