+++ Command/Handler.php
@@ -2,9 +2,9 @@
 
 namespace SumoCoders\FrameworkMultiUserBundle\Command;
 
-use SumoCoders\FrameworkMultiUserBundle\DataTransferObject\Interfaces\UserWithPasswordDataTransferObject;
+use SumoCoders\FrameworkMultiUserBundle\DataTransferObject\Interfaces\UserDataTransferObject;
 
 interface Handler
 {
-    public function handle(UserWithPasswordDataTransferObject $user);
+    public function handle(UserDataTransferObject $user);
 }
