+++ Console/DeleteUserCommand.php
@@ -3,7 +3,7 @@
 namespace SumoCoders\FrameworkMultiUserBundle\Console;
 
 use SumoCoders\FrameworkMultiUserBundle\Command\DeleteUserHandler;
-use SumoCoders\FrameworkMultiUserBundle\DataTransferObject\Interfaces\UserWithPasswordDataTransferObject;
+use SumoCoders\FrameworkMultiUserBundle\DataTransferObject\Interfaces\UserDataTransferObject;
 use SumoCoders\FrameworkMultiUserBundle\User\Interfaces\UserRepository;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
@@ -59,7 +59,7 @@ final class DeleteUserCommand extends Command
             return;
         }
 
-        $baseUserTransferObject = UserWithPasswordDataTransferObject::fromUser($user);
+        $baseUserTransferObject = UserDataTransferObject::fromUser($user);
 
         $this->handler->handle($baseUserTransferObject);
