+++ Tests/Command/CreateUserHandlerTest.php
@@ -3,11 +3,11 @@
 namespace SumoCoders\FrameworkMultiUserBundle\Tests\Command;
 
 use SumoCoders\FrameworkMultiUserBundle\Command\CreateUserHandler;
-use SumoCoders\FrameworkMultiUserBundle\DataTransferObject\UserWithPasswordDataTransferObject;
+use SumoCoders\FrameworkMultiUserBundle\DataTransferObject\UserDataTransferObject;
 use SumoCoders\FrameworkMultiUserBundle\User\InMemoryUserRepository;
 use SumoCoders\FrameworkMultiUserBundle\User\Interfaces\UserRepository;
 use SumoCoders\FrameworkMultiUserBundle\User\UserRepositoryCollection;
-use SumoCoders\FrameworkMultiUserBundle\User\UserWithPassword;
+use SumoCoders\FrameworkMultiUserBundle\User\User;
 use Symfony\Component\Security\Core\Encoder\EncoderFactory;
 use Symfony\Component\Security\Core\Encoder\PlaintextPasswordEncoder;
 
@@ -35,13 +35,13 @@ class CreateUserHandlerTest extends \PHPUnit_Framework_TestCase
     public function testCreateUserGetsHandled()
     {
         $handler = new CreateUserHandler(
-            new EncoderFactory([UserWithPassword::class => new PlaintextPasswordEncoder()]),
+            new EncoderFactory([User::class => new PlaintextPasswordEncoder()]),
             $this->userRepositoryCollection
         );
 
-        $user = new UserWithPassword('sumo', 'randomPassword', 'sumocoders', 'sumo@example.dev');
+        $user = new User('sumo', 'randomPassword', 'sumocoders', 'sumo@example.dev');
 
-        $userDataTransferObject = UserWithPasswordDataTransferObject::fromUser($user);
+        $userDataTransferObject = UserDataTransferObject::fromUser($user);
         $userDataTransferObject->plainPassword = 'randomPassword';
 
         $handler->handle($userDataTransferObject);
